+++ demo/js/site-ie8.js
@@ -1,4 +1,4 @@
-/*! formstone v1.2.1 [site-ie8.js] 2016-07-29 | GPL-3.0 License | formstone.it */
+/*! formstone v1.2.1 [site-ie8.js] 2016-08-02 | GPL-3.0 License | formstone.it */
 
 /**
 * @preserve HTML5 Shiv 3.7.3 | @afarkas @jdalton @jon_neal @rem | MIT/GPL2 Licensed