+++ src/js/tabs.js
@@ -1,7 +1,5 @@
 /* global define */
 
-// TODO: Fix vertical tabs
-
 (function(factory) {
 	if (typeof define === "function" && define.amd) {
 		define([
@@ -340,15 +338,13 @@
 			 * @param maxWidth [string] <Infinity> "Width at which to auto-disable plugin"
 			 * @param mobileMaxWidth [string] <'740px'> "Width at which to auto-disable mobile styles"
 			 * @param theme [string] <"fs-light"> "Theme class name"
-			 * @param vertical [boolean] <false> "Flag to draw vertical tab set"
 			 */
 
 			defaults: {
 				customClass       : "",
 				maxWidth          : Infinity,
 				mobileMaxWidth    : "740px",
-				theme             : "fs-light",
-				vertical          : false
+				theme             : "fs-light"
 			},
 
 			classes: [